pyproject.toml
[project]
name = "bilibili-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "bilibili-api-python>=17.4.0",
    "faster-whisper>=1.2.1",
    "litellm>=1.79.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
